refactor: change project structure to use modules
