+++ index.html
@@ -7,11 +7,11 @@
     <title>webix-practice</title>
     <script type="text/javascript" src="http://cdn.webix.com/edge/webix.js"></script>
     <link rel="stylesheet" type="text/css" href="http://cdn.webix.com/edge/webix.css">
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="./styles/main.css">
   </head>
   <body>
 
-    <script src="script.js" type="module" charset="utf-8"></script>
+    <script src="./scripts/app.js" type="module" charset="utf-8"></script>
 
   </body>
 </html>
