+++ scripts/section.main.js
@@ -1,0 +1,15 @@
+import { filmForm } from "./section.filmForm.js";
+import { sideMenu } from "./section.sideMenu.js";
+import { filmTable } from "./section.filmTable.js";
+
+const mainSection = { 
+  cols:[ 
+    sideMenu, 
+    { view:"resizer" },
+    filmTable, 
+    filmForm ]
+};
+
+export {
+  mainSection
+};
